requirements.txt
alembic==1.4.2
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
googlemaps==4.4.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
peewee==3.13.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.19
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
